refactored theme picker component
