+++ aio/src/app/shared/theme-picker/theme-picker.component.scss
