+++ aio/src/app/shared/theme-picker/theme-picker.component.spec.ts
@@ -1,0 +1,25 @@
+import {async, TestBed} from '@angular/core/testing';
+import { ThemePickerComponent } from './theme-picker.component';
+import { ThemeStorageService } from './theme-storage/theme-storage.service';
+
+
+describe('ThemePicker', () => {
+  let themeStorageService: any;
+
+  beforeEach(async(() => {
+    themeStorageService = jasmine.createSpyObj(['themeStorageService']);
+    themeStorageService.getStoredThemeName.and.returnValue(undefined);
+    TestBed.configureTestingModule({
+      declarations: [ThemePickerComponent],
+      providers: [ { provide: ThemeStorageService, useValue: themeStorageService } ]
+    }).compileComponents();
+  }));
+
+  it('should install theme based on name', () => {
+    const fixture = TestBed.createComponent(ThemePickerComponent);
+    const component = fixture.componentInstance;
+    const name = 'night-theme';
+    component.selectTheme(name);
+    expect(themeStorageService).toHaveBeenCalledWith(name);
+  });
+});
