+++ aio/src/app/shared/theme-picker/theme-storage/theme-storage.service.ts
@@ -5,25 +5,25 @@ export interface DocsSiteTheme {
   displayName?: string;
   accent: string;
   primary: string;
-  background: string;
+  isDark?: boolean;
   icon: string;
   isDefault?: boolean;
 }
 
 
 @Injectable()
-export class ThemeStorage {
+export class ThemeStorageService {
   static storageKey = 'docs-theme-storage-current-name';
 
   storeTheme(theme: DocsSiteTheme) {
     try {
-      window.localStorage[ThemeStorage.storageKey] = theme.name;
+      window.localStorage[ThemeStorageService.storageKey] = theme.name;
     } catch { }
   }
 
   getStoredThemeName(): string | null {
     try {
-      return window.localStorage[ThemeStorage.storageKey] || null;
+      return window.localStorage[ThemeStorageService.storageKey] || null;
     } catch {
       return null;
     }
@@ -31,7 +31,7 @@ export class ThemeStorage {
 
   clearStorage() {
     try {
-      window.localStorage.removeItem(ThemeStorage.storageKey);
+      window.localStorage.removeItem(ThemeStorageService.storageKey);
     } catch { }
   }
 }
